+++ EmpowerPlant.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -15,10 +15,10 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/getsentry/sentry-cocoa",
       "state" : {
-        "revision" : "21223d1c864db0561d91f48d80f269a363a1625d",
-        "version" : "8.47.0"
+        "revision" : "930b78a63f47549c81e6e63c9172584f7d3dfdd6",
+        "version" : "8.52.1"
       }
     }
   ],
-  "version" : 2
+  "version" : 3
 }
