+++ EmpowerPlant/Product+CoreDataProperties.swift
@@ -12,18 +12,4 @@ import CoreData
 
 extension Product {
 
-    @nonobjc public class func fetchRequest() -> NSFetchRequest<Product> {
-        let fr = NSFetchRequest<Product>(entityName: "Product")
-        fr.sortDescriptors = [.init(key: "title", ascending: true)]
-        return fr
-    }
-    
-//    @NSManaged public var text: String?
-    @NSManaged public var title: String?
-    @NSManaged public var productDescription: String?
-    @NSManaged public var productDescriptionFull: String?
-    @NSManaged public var productId: String?
-    @NSManaged public var img: String?
-    @NSManaged public var imgCropped: String?
-    @NSManaged public var price: String?
 }
